+++ sdk/notificationhubs/notification-hubs/src/api/index.ts
@@ -143,6 +143,7 @@ export * from "./getNotificationOutcomeDetails.js";
 export * from "./getRegistration.js";
 export * from "./listNotificationHubJobs.js";
 export * from "./listRegistrations.js";
+export * from "./listRegistrationsByDevice.js";
 export * from "./listRegistrationsByTag.js";
 export * from "./scheduleNotification.js";
 export * from "./sendNotification.js";